Layer feature (#2)
* Correct stream positioning

* Layer chunk class

* Layer merging working
